remove pipe operator
This is the story of a little Olea operator spelled `|>`. It was a postfix operator that affected precedence similar to Haskell's `$`. It would basically group the entire left expression, e.g. `x + y |> * z` would parse as `(x + y) * z`, and `-foo |> .to_string()` would parse as `(-foo).to_string()`. I have yet to find a use for this operator in practice, perhaps in part because so many unary operators are already postfix in Olea, or perhaps because it's poorly conceived in the first place. The parsing implementation failed when a postfix operator followed the pipe, and it was much more work to fix it than remove the pipe altogether. See you in another life, pipe.