fix arborist colon bug
Previously, if a colon was immediately followed by its body with no newline in between, subsequent items on the same indentation level would be erroneously included in the block. For example, this Olea code...
```
a: 1
b: 2
```
... was arborized as `a {1; b {2}}`. Now, it correctly parses as `a {1}; b {2}`.

There is an additional bug that prevents writing trailing commas in the function call in `example.olea`. This will be more difficult to fix, so for now I won't! :^)