fix field access semantics
Struct field access now operate on structs rather than pointers to structs. As far as I can tell, we now match Rust's semantics, besides implicit dereferencing, which is less necessary because Olea uses postfix deref syntax.

If the operand is a place expression, we get a pointer to its memory region. Otherwise, we make a stack allocation to store the value and get a pointer to that. This is the same semantics as the ref `@` operator. We should come up with a good name for this operation! After that we do the pointer offset as before. Field access is itself a place expression, so if it is used as a value expression then this pointer will be dereferenced.