fix register allocation soundness bug from unused registers (oops!)
This bug was a proper algorithmic error! Not some typo, or some old code that doesn't match a refactor.

For register allocation, the process of choosing which machine registers should store the value of each IR register. The basic principle is that if two IR registers can be "live" (have its value used at some point) at the same time, they must be stored in separate machine registers.

The bug arises from how we treat registers that are never used by another instruction. We consider these to have an empty lifetime, which in and of itself is fine. However, this also means that such a register was never considered "live" at the same time as any other register, meaning the allocator erroneously thought they could share a physical register. In practice, all unused registers would be allocated to fox32's `r0`, corrupting any other register stored there if it happened to be live.

We fix this with a simple pre-pass that allocates these unused registers to the "dummy" physical register `r31`, which we use for temporaries that do not need to live across IR instructions.